Added sorting in a grid
